pyproject.toml
[project]
name = "Pure-Python-Geospatial-Export"
version = "0.1.0"
dependencies = [
    "geomet==1.1.0",
    "geopackage==1.0.0",
    "pyshp==2.3.1",
]

[project.optional-dependencies]
gis = [
    "geopandas==1.0.1",
    "pandas==2.3.0",
    "shapely==2.0.7",
]
snowflake = [
    "snowflake-snowpark-python==1.20.0"
]
